refactor: model 패키지 이동
