+++ backend/src/main/java/kernel360/techpick/feature/structure/service/node/server/FolderServerNode.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import jakarta.validation.constraints.NotNull;
+import kernel360.techpick.feature.structure.model.NameProvider;
 import kernel360.techpick.feature.structure.service.node.client.ClientNode;
 import kernel360.techpick.feature.structure.service.node.client.FolderClientNode;
-import kernel360.techpick.feature.structure.service.model.NameProvider;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
